pyproject.toml
[project]
name = "larrak2"
version = "0.1.0"
description = "Multi-objective Pareto optimization for thermo-gear systems"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pymoo>=0.6",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy-stl>=3.0",
    "torch>=2.0",
    "scikit-learn>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1", "mypy>=1.0"]
casadi = ["casadi>=3.6"]
deap = ["deap>=1.3"]
orchestration = ["redis>=5.0", "weaviate-client>=4.0"]

[project.scripts]
larrak-run = "larrak2.cli.run:main"
larrak-gen = "larrak2.cli.generate:main"
larrak-train = "larrak2.cli.train:main"
larrak-validate = "larrak2.cli:validate_main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests/ci"]
norecursedirs = ["tests/dev"]
python_files = "test_*.py"
